MOD: Rewrite Command Controller to cmd+bias arch, when +- or <> CMD has current bias value which can help up improve performance of the machine RLE compression workes when source is loaded.
FIX: CyclePush/Pop Stack error - when we return to cycle pegin we need to Pop IP, and Push it again.

FIX: char is signed, so algorithm works incorrect. Change to Uint8_t for any cmd and bias values

ADD: some samples, but they are not working. Have some problems with current CPU implementation